refactor: 출석 상태를 판단하는 메서드를 StandardTime 내부로 이동
