+++ src/main/java/domain/AttendanceStatus.java
@@ -1,8 +1,5 @@
 package domain;
 
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-
 public enum AttendanceStatus {
     ATTENDANCE("출석"),
     LATENESS("지각"),
@@ -12,17 +9,6 @@ public enum AttendanceStatus {
 
     AttendanceStatus(String name) {
         this.name = name;
-    }
-
-    public static AttendanceStatus judge(LocalDateTime localDateTime) {
-        LocalTime attendanceTime = localDateTime.toLocalTime();
-        StandardTime standardTime = StandardTime.findByDayOfWeek(localDateTime.toLocalDate().getDayOfWeek());
-
-        if (attendanceTime.isAfter(standardTime.getAbsentTime())) return AttendanceStatus.ABSENCE;
-
-        if (attendanceTime.isAfter(standardTime.getLateTime())) return AttendanceStatus.LATENESS;
-
-        return AttendanceStatus.ATTENDANCE;
     }
 
     public String getName() {
