+++ src/test/java/domain/StandardTimeTest.java
@@ -1,10 +1,13 @@
 package domain;
 
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.time.DayOfWeek;
+import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.stream.Stream;
 
@@ -35,5 +38,95 @@ class StandardTimeTest {
 
         assertThat(standardTime.getLateTime()).isEqualTo(lateTime);
         assertThat(standardTime.getAbsentTime()).isEqualTo(absentTime);
+    }
+
+    @Test
+    void 월요일_출석_테스트() {
+        LocalDateTime localDateTime = LocalDateTime.of(2024, 12, 16, 13, 0);
+
+        AttendanceStatus attendanceStatus = StandardTime.judge(localDateTime);
+
+        Assertions.assertThat(attendanceStatus).isEqualTo(AttendanceStatus.ATTENDANCE);
+    }
+
+    @Test
+    void 월요일_출석_경계값_테스트() {
+        LocalDateTime localDateTime = LocalDateTime.of(2024, 12, 16, 13, 5);
+
+        AttendanceStatus attendanceStatus = StandardTime.judge(localDateTime);
+
+        Assertions.assertThat(attendanceStatus).isEqualTo(AttendanceStatus.ATTENDANCE);
+    }
+
+    @Test
+    void 월요일_지각_테스트() {
+        LocalDateTime localDateTime = LocalDateTime.of(2024, 12, 16, 13, 6);
+
+        AttendanceStatus attendanceStatus = StandardTime.judge(localDateTime);
+
+        Assertions.assertThat(attendanceStatus).isEqualTo(AttendanceStatus.LATENESS);
+    }
+
+    @Test
+    void 월요일_지각_경계값_테스트() {
+        LocalDateTime localDateTime = LocalDateTime.of(2024, 12, 16, 13, 30);
+
+        AttendanceStatus attendanceStatus = StandardTime.judge(localDateTime);
+
+        Assertions.assertThat(attendanceStatus).isEqualTo(AttendanceStatus.LATENESS);
+    }
+
+    @Test
+    void 월요일_결석_테스트() {
+        LocalDateTime localDateTime = LocalDateTime.of(2024, 12, 16, 13, 31);
+
+        AttendanceStatus attendanceStatus = StandardTime.judge(localDateTime);
+
+        Assertions.assertThat(attendanceStatus).isEqualTo(AttendanceStatus.ABSENCE);
+    }
+
+    @Test
+    void 화요일_출석_테스트() {
+        LocalDateTime localDateTime = LocalDateTime.of(2024, 12, 17, 10, 0);
+
+        AttendanceStatus attendanceStatus = StandardTime.judge(localDateTime);
+
+        Assertions.assertThat(attendanceStatus).isEqualTo(AttendanceStatus.ATTENDANCE);
+    }
+
+    @Test
+    void 화요일_출석_경계값_테스트() {
+        LocalDateTime localDateTime = LocalDateTime.of(2024, 12, 17, 10, 5);
+
+        AttendanceStatus attendanceStatus = StandardTime.judge(localDateTime);
+
+        Assertions.assertThat(attendanceStatus).isEqualTo(AttendanceStatus.ATTENDANCE);
+    }
+
+    @Test
+    void 화요일_지각_테스트() {
+        LocalDateTime localDateTime = LocalDateTime.of(2024, 12, 17, 10, 6);
+
+        AttendanceStatus attendanceStatus = StandardTime.judge(localDateTime);
+
+        Assertions.assertThat(attendanceStatus).isEqualTo(AttendanceStatus.LATENESS);
+    }
+
+    @Test
+    void 화요일_지각_경계값_테스트() {
+        LocalDateTime localDateTime = LocalDateTime.of(2024, 12, 17, 10, 30);
+
+        AttendanceStatus attendanceStatus = StandardTime.judge(localDateTime);
+
+        Assertions.assertThat(attendanceStatus).isEqualTo(AttendanceStatus.LATENESS);
+    }
+
+    @Test
+    void 화요일_결석_테스트() {
+        LocalDateTime localDateTime = LocalDateTime.of(2024, 12, 17, 10, 31);
+
+        AttendanceStatus attendanceStatus = StandardTime.judge(localDateTime);
+
+        Assertions.assertThat(attendanceStatus).isEqualTo(AttendanceStatus.ABSENCE);
     }
 }
